Added performance patch
The performance patch updates a small section of the vector field. This approach however introduces ghosting of past cells on a moving emitter which is fixed by allowing a variable obx size that depends on how much did the emitter bounding box move in one frame. This allows to have always the least amount of updates while keeping fidelity and preventing ghosting/smearing